Restore Python implementation
