+++ src/mapfile_parser/mapfile.py
@@ -13,163 +13,15 @@ from pathlib import Path
 from .progress_stats import ProgressStats
 from . import utils
 
-from .mapfile_parser import FoundSymbolInfo as FoundSymbolInfo
-from .mapfile_parser import SymbolComparisonInfo as SymbolComparisonInfo
-from .mapfile_parser import MapsComparisonInfo as MapsComparisonInfo
-from .mapfile_parser import Symbol as Symbol
-from .mapfile_parser import File as File
-from .mapfile_parser import Segment as Segment
-from .mapfile_parser import MapFile as MapFile
+
+regex_fileDataEntry = re.compile(r"^\s+(?P<section>\.[^\s]+)\s+(?P<vram>0x[^\s]+)\s+(?P<size>0x[^\s]+)\s+(?P<name>[^\s]+)$")
+regex_functionEntry = re.compile(r"^\s+(?P<vram>0x[^\s]+)\s+(?P<name>[^\s]+)$")
+# regex_functionEntry = re.compile(r"^\s+(?P<vram>0x[^\s]+)\s+(?P<name>[^\s]+)((\s*=\s*(?P<expression>.+))?)$")
+regex_label = re.compile(r"^(?P<name>\.?L[0-9A-F]{8})$")
+regex_fill = re.compile(r"^\s+(?P<fill>\*[^\s\*]+\*)\s+(?P<vram>0x[^\s]+)\s+(?P<size>0x[^\s]+)\s*$")
+regex_segmentEntry = re.compile(r"(?P<name>([^\s]+)?)\s+(?P<vram>0x[^\s]+)\s+(?P<size>0x[^\s]+)\s+(?P<loadaddress>(load address)?)\s+(?P<vrom>0x[^\s]+)$")
 
 
-def __symbolrs_serializeVram(self: Symbol, humanReadable: bool=True) -> str|int|None:
-    if humanReadable:
-        return f"0x{self.vram:08X}"
-    return self.vram
-
-def __symbolrs_serializeSize(self: Symbol, humanReadable: bool=True) -> str|int|None:
-    if self.size is None:
-        return None
-    if humanReadable:
-        return f"0x{self.size:X}"
-    return self.size
-
-def __symbolrs_serializeVrom(self: Symbol, humanReadable: bool=True) -> str|int|None:
-    if self.vrom is None:
-        return None
-    if humanReadable:
-        return f"0x{self.vrom:06X}"
-    return self.vrom
-
-def __symbolrs_toJson(self: Symbol, humanReadable: bool=True) -> dict[str, Any]:
-    result: dict[str, Any] = {
-        "name": self.name,
-        "vram": self.serializeVram(humanReadable=humanReadable),
-        "size": self.serializeSize(humanReadable=humanReadable),
-        "vrom": self.serializeVrom(humanReadable=humanReadable),
-    }
-    return result
-
-Symbol.serializeVram = __symbolrs_serializeVram
-Symbol.serializeSize = __symbolrs_serializeSize
-Symbol.serializeVrom = __symbolrs_serializeVrom
-Symbol.toJson = __symbolrs_toJson
-
-
-@property
-def __filers_filepath(self: File) -> Path:
-    return Path(self._filepath_internal)
-@__filers_filepath.setter
-def __filers_filepath(self: File, newPath: Path):
-    self._filepath_internal = str(newPath)
-
-def __filers_serializeVram(self: File, humanReadable: bool=True) -> str|int|None:
-    if humanReadable:
-        return f"0x{self.vram:08X}"
-    return self.vram
-
-def __filers_serializeSize(self: File, humanReadable: bool=True) -> str|int|None:
-    if humanReadable:
-        return f"0x{self.size:X}"
-    return self.size
-
-def __filers_serializeVrom(self: File, humanReadable: bool=True) -> str|int|None:
-    if self.vrom is None:
-        return None
-    if humanReadable:
-        return f"0x{self.vrom:06X}"
-    return self.vrom
-
-def __filers_toJson(self: File, humanReadable: bool=True) -> dict[str, Any]:
-    fileDict: dict[str, Any] = {
-        "filepath": str(self.filepath),
-        "sectionType": self.sectionType,
-        "vram": self.serializeVram(humanReadable=humanReadable),
-        "size": self.serializeSize(humanReadable=humanReadable),
-        "vrom": self.serializeVrom(humanReadable=humanReadable),
-    }
-
-    symbolsList = []
-    for symbol in self:
-        symbolsList.append(symbol.toJson(humanReadable=humanReadable))
-
-    fileDict["symbols"] = symbolsList
-    return fileDict
-
-File.filepath = __filers_filepath
-File.serializeVram = __filers_serializeVram
-File.serializeSize = __filers_serializeSize
-File.serializeVrom = __filers_serializeVrom
-File.toJson = __filers_toJson
-
-
-def __segmentrs_serializeVram(self: Segment, humanReadable: bool=True) -> str|int|None:
-    if humanReadable:
-        return f"0x{self.vram:08X}"
-    return self.vram
-
-def __segmentrs_serializeSize(self: Segment, humanReadable: bool=True) -> str|int|None:
-    if humanReadable:
-        return f"0x{self.size:X}"
-    return self.size
-
-def __segmentrs_serializeVrom(self: Segment, humanReadable: bool=True) -> str|int|None:
-    if humanReadable:
-        return f"0x{self.vrom:06X}"
-    return self.vrom
-
-def __segmentrs_toJson(self: Segment, humanReadable: bool=True) -> dict[str, Any]:
-    segmentDict: dict[str, Any] = {
-        "name": self.name,
-        "vram": self.serializeVram(humanReadable=humanReadable),
-        "size": self.serializeSize(humanReadable=humanReadable),
-        "vrom": self.serializeVrom(humanReadable=humanReadable),
-    }
-
-    filesList = []
-    for file in self:
-        filesList.append(file.toJson(humanReadable=humanReadable))
-
-    segmentDict["files"] = filesList
-
-    return segmentDict
-
-Segment.serializeVram = __segmentrs_serializeVram
-Segment.serializeSize = __segmentrs_serializeSize
-Segment.serializeVrom = __segmentrs_serializeVrom
-Segment.toJson = __segmentrs_toJson
-
-
-
-def __mapfilers_printAsCsv(self: MapFile, printVram: bool=True, skipWithoutSymbols: bool=True):
-    print(self.toCsv(printVram=printVram, skipWithoutSymbols=skipWithoutSymbols), end="")
-
-def __mapfilers_printSymbolsCsv(self: MapFile):
-    print(self.toCsvSymbols(), end="")
-
-def __mapfilers_toJson(self: MapFile, humanReadable: bool=True) -> dict[str, Any]:
-    segmentsList = []
-    for segment in self:
-        segmentsList.append(segment.toJson(humanReadable=humanReadable))
-
-    result: dict[str, Any] = {
-        "segments": segmentsList
-    }
-    return result
-
-MapFile.printAsCsv = __mapfilers_printAsCsv
-MapFile.printSymbolsCsv = __mapfilers_printSymbolsCsv
-MapFile.toJson = __mapfilers_toJson
-
-
-# regex_fileDataEntry = re.compile(r"^\s+(?P<section>\.[^\s]+)\s+(?P<vram>0x[^\s]+)\s+(?P<size>0x[^\s]+)\s+(?P<name>[^\s]+)$")
-# regex_functionEntry = re.compile(r"^\s+(?P<vram>0x[^\s]+)\s+(?P<name>[^\s]+)$")
-# # regex_functionEntry = re.compile(r"^\s+(?P<vram>0x[^\s]+)\s+(?P<name>[^\s]+)((\s*=\s*(?P<expression>.+))?)$")
-# regex_label = re.compile(r"^(?P<name>\.?L[0-9A-F]{8})$")
-# regex_fill = re.compile(r"^\s+(?P<fill>\*[^\s\*]+\*)\s+(?P<vram>0x[^\s]+)\s+(?P<size>0x[^\s]+)\s*$")
-# regex_segmentEntry = re.compile(r"(?P<name>([^\s]+)?)\s+(?P<vram>0x[^\s]+)\s+(?P<size>0x[^\s]+)\s+(?P<loadaddress>(load address)?)\s+(?P<vrom>0x[^\s]+)$")
-
-"""
 @dataclasses.dataclass
 class FoundSymbolInfo:
     file: File
@@ -190,9 +42,7 @@ class FoundSymbolInfo:
             message = "Symbol"
         return f"{message} {self.getAsStr()}"
 
-"""
 
-"""
 @dataclasses.dataclass
 class SymbolComparisonInfo:
     symbol: Symbol
@@ -207,9 +57,8 @@ class MapsComparisonInfo:
         self.badFiles: set[File] = set()
         self.missingFiles: set[File] = set()
         self.comparedList: list[SymbolComparisonInfo] = []
-"""
 
-"""
+
 @dataclasses.dataclass
 class Symbol:
     name: str
@@ -284,9 +133,8 @@ class Symbol:
     def __hash__(self):
         return hash((self.name, self.vram))
 
-"""
 
-"""
+
 @dataclasses.dataclass
 class File:
     filepath: Path
@@ -419,6 +267,19 @@ class File:
         return fileDict
 
 
+    def copySymbolList(self) -> list[Symbol]:
+        """Returns a copy (not a reference) of the internal symbol list"""
+        return list(self._symbols)
+
+    def setSymbolList(self, newList: list[Symbol]) -> None:
+        """Replaces the internal symbol list with a copy of `newList`"""
+        self._symbols = list(newList)
+
+    def appendSymbol(self, sym: Symbol) -> None:
+        """Appends a copy of `sym` into the internal symbol list"""
+        self._symbols.append(sym)
+
+
     def __iter__(self) -> Generator[Symbol, None, None]:
         for sym in self._symbols:
             yield sym
@@ -440,9 +301,8 @@ class File:
     # https://stackoverflow.com/a/56915493/6292472
     def __hash__(self):
         return hash((self.filepath,))
-"""
 
-"""
+
 @dataclasses.dataclass
 class Segment:
     name: str
@@ -580,6 +440,19 @@ class Segment:
         return segmentDict
 
 
+    def copyFileList(self) -> list[File]:
+        """Returns a copy (not a reference) of the internal file list"""
+        return list(self._filesList)
+
+    def setFileList(self, newList: list[File]) -> None:
+        """Replaces the internal file list with a copy of `newList`"""
+        self._filesList = list(newList)
+
+    def appendFile(self, file: File) -> None:
+        """Appends a copy of `file` into the internal file list"""
+        self._filesList.append(file)
+
+
     def __iter__(self) -> Generator[File, None, None]:
         for file in self._filesList:
             yield file
@@ -598,26 +471,16 @@ class Segment:
     # https://stackoverflow.com/a/56915493/6292472
     def __hash__(self):
         return hash((self.name, self.vram, self.size, self.vrom))
-"""
 
-"""
+
 class MapFile:
     def __init__(self):
         self._segmentsList: list[Segment] = list()
-        self._internalMap = MapFileRs()
+
+        # Deprecated
         self.debugging: bool = False
-        self._originalParser = False
 
     def readMapFile(self, mapPath: Path):
-        self._internalMap.readMapFile(mapPath)
-        # for segment in tempRs.segments_list:
-        #     print(segment.name)
-        #     continue
-        #     for file in segment.files_list:
-        #         for symbol in segment.symbols:
-        #             pass
-        if not self._originalParser:
-            return
         with mapPath.open("r") as f:
             mapData = f.read()
 
@@ -970,6 +833,19 @@ class MapFile:
         return result
 
 
+    def copySegmentList(self) -> list[Segment]:
+        """Returns a copy (not a reference) of the internal segment list"""
+        return list(self._segmentsList)
+
+    def setSegmentList(self, newList: list[Segment]) -> None:
+        """Replaces the internal segment list with a copy of `newList`"""
+        self._segmentsList = list(newList)
+
+    def appendSegment(self, segment: Segment) -> None:
+        """Appends a copy of `segment` into the internal segment list"""
+        self._segmentsList.append(segment)
+
+
     def __iter__(self) -> Generator[Segment, None, None]:
         for file in self._segmentsList:
             yield file
@@ -979,4 +855,3 @@ class MapFile:
 
     def __len__(self) -> int:
         return len(self._segmentsList)
-"""
