+++ src/rs/file.rs
@@ -13,7 +13,7 @@ use pyo3::prelude::*;
 use pyo3::class::basic::CompareOp;
 
 #[derive(Debug, Clone)]
-#[pyclass(module = "mapfile_parser", unsendable, sequence)]
+#[pyclass(module = "mapfile_parser", sequence)]
 pub struct File {
     #[pyo3(get, set, name="_filepath_internal")]
     // TODO: pyo3 exposes this as str, need to fix somehow
